feat: 설정 화면 정비 — 외부 연결 변경 및 사용법 가이드 분리
- 문의하기: 이메일 → Instagram(@mordecai_kwan)
- 개인정보 처리방침: Notion 페이지로 교체
- 앱 평가: StoreKit 인앱 시트 → App Store 리뷰 작성 페이지 직접 연결
- 사용법 가이드: 나열되던 항목을 NavigationLink 한 줄로 축소, 별도 화면(FeatureGuideView)으로 분리
